requirements.txt
# Requirements for UAV Deconfliction System
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0
scipy>=1.7.0
pytest>=6.2.0
pyyaml>=5.4.0
